Refactor out MerchantProfile from MerchantSideBar
